krust
Module init

Initialization of Cortex-M device

Cortex-M devices are using a vector table for initialization. It is located at the start of the code region in memory.

The first 2 elements of the vector table are :

the initial Stack Pointer

the Reset vector (pointer to the Reset handler)

This crates takes care of RAM initialization, and populates the vector table like this :

Vector Table
+------------------+ 0x0
| Initial SP       |
+------------------+ 0x4
| (1) Reset        |
+------------------+ 0x8
| (2) NMI          |
+------------------+ 0xc
| (3) HardFault    |
+------------------+ 0x10
| (4) MemManage    |
+------------------+ 0x14
| (5) BusFault     |
+------------------+ 0x18
| (6) UsageFault   |
+------------------+ 0x1c
| (7-10) Reserved  |
+------------------+ 0x2c
| (11) SVCall      |
+------------------+ 0x30
| (12-13) Reserved |
+------------------+ 0x38
| (14) PendSV      |
+------------------+ 0x3c
| (15) SysTick     |
+------------------+ 0x40
| IRQ_n            |
| ...              |

Re-exports

pub use crate::init::systick::start_sys_tick;

Modules

handlers 🔒

panic 🔒

systick 🔒

Structs

ExceptionsHandlers

Statics

RESET_VECTOR

Reset vector, part of the Vector table, points to our Reset handler : Reset()

_EXCEPTIONS

Exceptions vectors, part of the Vector table

Functions

Reset

Our Reset handler, wich initializes RAM and calls main

enable_system_handler_fault⚠

init_bss 🔒 ⚠

Initialization of the .bss section by zeroing out memory

init_data 🔒 ⚠

Initialization of the .data section by populating it with data from sidata