krust::proc
Struct SystemProcess

pub struct SystemProcess {
    last_proc_id: u16,
    process_list: LinkedList<Process>,
    current_process_id: u16,
    current_mpu_conf: Option<Mpu>,
}

This struct hold reference to the Process List of the system, and the PID of the running process

All operation performed on process are implemented here (create, kill, schedule, …)

Fields
last_proc_id: u16process_list: LinkedList<Process>current_process_id: u16current_mpu_conf: Option<Mpu>
Implementations

impl SystemProcess

pub fn new() -> SystemProcess

fn get_new_proc_id(&mut self) -> u16

Get free PID for new process

pub fn get_process_by_id(&mut self, proc_id: u16) -> Option<Process>

Get non mutable reference of a process from a PID

pub fn get_current_process(&mut self) -> Option<Process>

Get non mutable reference of the running process

pub fn get_current_priority_process(&mut self) -> u8

Get priority of the running process

pub fn create_process( &mut self, name: &'static str, code_ptr: &[u8], code_len: usize, priority: u8, ) -> u16

Creates a new process, assigns a process ID, allocates memory for its code and stack, initializes the stack frame, and adds the new process to the process list.

Arguments

name - A string representing the name of the new process.

code_ptr - A byte slice containing the code to be loaded into the process.

code_len - The length of the code to be loaded.

Returns

A unique process ID (PID) for the newly created process.

IMPORTANT

Process code MUST end with a SYS_EXIT then an infinite loop

fn load_process_code(&mut self, code_ptr: &[u8], code_len: usize) -> *mut u8

Loads the code for a new process into heap memory and returns the pointer to the allocated space.

Arguments

code_ptr - A byte slice containing the process code to be loaded.

code_len - The length of the code to be loaded.

Returns

A pointer to the allocated memory containing the process code.

fn create_init_stack_frame(&mut self, stack_ptr: *mut u8, entry_point: *mut u8)

Initializes the stack frame for a new process. This function sets up the initial values on the stack, such as the entry point and xPSR value, which are required when the process starts execution.

The 14 first u32 (save of register r0 to r12 + LR) are set to 0x00 so that new process starts with clean registers Then we set up the entry point and the xPSR.

When the scheduler will pick up a process for the fisrt time, it’ll unwrap the INIT STACK, setting up all registers properly

INIT STACK

+--------+ < SP
| 0x00   |
+--------+
| ...    |
+--------+ < SP + 14 * REG_SIZE
| PC     |
+--------+ < SP + 15 * REG_SIZE
| xPSR   |
+--------+

pub fn kill_process(&mut self, proc_id: u16)

Kill a specific process based on a PID

pub fn exit_current_process(&mut self)

Mark the running process as Finished, so that this process get killed on next scheduler call

pub fn list_proc(&mut self)

List process in the Process List with the following format : [PID] PROC_NAME (PROC_STATUS)

pub fn enable_current_mpu(&self)

pub fn disable_current_mpu(&self)

fn mpu_region_size_from_memory_len(&self, size: usize) -> u32

fn next_power_of_two_exponent(&self, n: usize) -> usize

pub fn schedule_next_process(&mut self)

Schedules the next process to run. This function also handles killing processes that have finished execution.

It performs the following:

Collects the processes that are finished and calls kill_process on them.

Iterates over the list of processes to identify the next idle process and the currently running process based on priority.

Marks the current running process as idle, saves its state, and schedules the next process.

Panics

This function will panic with the message "NOTHING TO DO" if it can’t find a next process to schedule.